+++ backend/handlers/math/market/marketvolume.go
@@ -3,13 +3,13 @@ package marketmath
 import "socialpredict/models"
 
 // getMarketVolume returns the total volume of trades for a given market
-func GetMarketVolume(bets []models.Bet) uint {
-	var totalVolume float64
+func GetMarketVolume(bets []models.Bet) int64 {
+	var totalVolume int64
 	for _, bet := range bets {
 		totalVolume += bet.Amount
 	}
 
-	totalVolumeUint := uint(totalVolume)
+	totalVolumeUint := int64(totalVolume)
 
 	return totalVolumeUint
 }
